+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🐞 Fixed
 - Fix channel data being overridden by events with stale channel data [#4278](https://github.com/GetStream/stream-chat-swift/pull/4278)
+- Fix UI hang caused by activating or deactivating the audio session on the main thread [#4286](https://github.com/GetStream/stream-chat-swift/pull/4286)
 
 ### 🔄 Changed