+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## StreamChatUI
 ### 🐞 Fixed
+- Fix missing paddings on the iPhone Duo [#4287](https://github.com/GetStream/stream-chat-swift/pull/4287)
+- Fix the message actions pop-up not being shown on the iPhone Duo [#4287](https://github.com/GetStream/stream-chat-swift/pull/4287)
+- Fix a crash when the gallery is open on the iPhone Duo [#4287](https://github.com/GetStream/stream-chat-swift/pull/4287)
+- Fix the composer caret not being aligned with the placeholder on the iPhone Duo [#4287](https://github.com/GetStream/stream-chat-swift/pull/4287)
 - Fix incorrect rounded corner on image attachments shown below other attachments [#4277](https://github.com/GetStream/stream-chat-swift/pull/4277)
 - Fix empty voice recording preview when quoting a message with multiple attachments [#4276](https://github.com/GetStream/stream-chat-swift/pull/4276)